requirements.txt
BeautifulSoup==3.2.1
Django==1.5.4
Fabric==1.8.3
PyYAML==3.10
South==0.8.4
Unidecode==0.04.14
Werkzeug==0.9.4
argparse==1.1
django-ace==1.0.1
django-admintheme==0.2
django-annoying==0.7.9
django-autofixture==0.6.3
django-countries==1.5
django-debug-toolbar==1.0.1
django-email-extras==0.1.13
django-extensions==1.3.3
django-forms-builder==0.9
-e git://github.com/Soaa-/django-nested-inlines.git@aaf995a223720ca7d6b4bfc74fb5707cd1bd08dd#egg=django_nested_inlines-master
django-pagination==1.0.7
django-test-utils==0.3
djorm-ext-pgtrgm==0.1
ecdsa==0.11
gevent==1.0
greenlet==0.4.2
paramiko==1.12.3
preggy==0.11.1
psycopg2==2.5.2
py-mysql2pgsql==0.1.5
pyVows==2.0.4
pycrypto==2.6.1
pygraphviz==1.2
python-gnupg==0.3.6
requests==2.0.1
six==1.5.2
sphinx-me==0.2.1
sqlparse==0.1.11
termcolor==1.1.0
wsgiref==0.1.2
